+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 9.07.2018
 
+* [Rust] Code Royale - Handle queen refactoring (CodinGame)
 * [Rust] Code Royale - Queen escape Game method (CodinGame)
 
 #### 8.07.2018